Add linter config
When considering including this project as portfolio material I took a
look at this almost four year old code and immediately was disappointed
in myself. I've grown way too much to make the current codebase a
profile-pinned repo. To start with this needs a linter and matching
config pronto.

Also how did this app not have a lock file in the repo? That's crazy.